chore(next): refactor components folder
