+++ packages/next/src/client/components/actions/SignOutButton.tsx
@@ -19,12 +19,12 @@
 'use client';
 
 import {FC, forwardRef, HTMLAttributes, PropsWithChildren, ReactElement, Ref} from 'react';
-import InternalAuthAPIRoutesConfig from '../../configs/InternalAuthAPIRoutesConfig';
+import InternalAuthAPIRoutesConfig from '../../../configs/InternalAuthAPIRoutesConfig';
 
 /**
- * Props interface of {@link SignInButton}
+ * Interface for SignInButton component props.
  */
-export type SignInButtonProps = HTMLAttributes<HTMLButtonElement>;
+export type SignOutButtonProps = HTMLAttributes<HTMLButtonElement>;
 
 /**
  * SignInButton component. This button initiates the sign-in process when clicked.
@@ -43,15 +43,15 @@ export type SignInButtonProps = HTMLAttributes<HTMLButtonElement>;
  * }
  * ```
  */
-const SignInButton: FC<PropsWithChildren<SignInButtonProps>> = forwardRef<
+const SignOutButton: FC<PropsWithChildren<SignOutButtonProps>> = forwardRef<
   HTMLButtonElement,
-  PropsWithChildren<SignInButtonProps>
+  PropsWithChildren<SignOutButtonProps>
 >(
   (
-    {children = 'Sign In', className, style, ...rest}: PropsWithChildren<SignInButtonProps>,
+    {children = 'Sign Out', className, style, ...rest}: PropsWithChildren<SignOutButtonProps>,
     ref: Ref<HTMLButtonElement>,
   ): ReactElement => (
-    <form action={InternalAuthAPIRoutesConfig.signIn}>
+    <form action={InternalAuthAPIRoutesConfig.signOut}>
       <button ref={ref} className={className} style={style} type="submit" {...rest}>
         {children}
       </button>
@@ -59,4 +59,4 @@ const SignInButton: FC<PropsWithChildren<SignInButtonProps>> = forwardRef<
   ),
 );
 
-export default SignInButton;
+export default SignOutButton;
